Added composite key references
Added an additional mechanism by which regulatory bodies, units, counted items, quantities, dimensions and prefixes can be referenced within an instance document using composite keys. This is further restricted by applying uniqueness constraints and requiring the histories of these elements to be strictly linear with each element being derived from an explicit ancestor or constituting the initial version of the element. Furthermore, the concept of “domains” is used to associate elements with a specific regulatory body or entity. Timestamps were added to supplement the versioning of elements. References using a URL is still permitted to allow for some backward compatibility requiring only minor changes.
Because the root units, as defined by the enums, neither have a version history nor an explicit definition (same goes for the prefix enum), the derivation of units was extended beyond the concept of root units by allowing references to normal Unit elements. A unit element without root units or references specified can implicitly be regarded as a base or root unit.