Added ConstantSet
Added ConstantSet and Constant elements to allow for the definition of scalar real and imaginary constants which may reference units or counted items but not quantities, because constants may be used for multiple quantities such as pi, e, electron charge and others. TODO: Possibly add constant vectors, tensors etc. The simplistic implementation presented here should be discussed to further improve the applicability.